requirements.txt
# Educational RLHF Real-time System - Complete Project

## Requirements
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-generativeai==0.8.5
numpy==1.26.2
python-dotenv==1.0.0
asyncio==3.4.3
loguru==0.7.2